Split out batch
