+++ lib/iterator/batch_test.go
